+++ packages/core/alfa-core/stories/demos/basic.tsx
@@ -1,20 +1,29 @@
 import React, { useRef, useEffect } from 'react';
-import createMicroApp from '../../src';
+import { createMicroApp } from '../../src';
 import { appManifest } from './constants';
 
 const Basic: React.FC<{}> = () => {
   const appRef = useRef();
+
   useEffect(() => {
     (async () => {
+
       const app = await createMicroApp({
-        container: appRef.current,
         ...appManifest
       }, {})
+
       await app.load()
-      //@ts-ignore
-      await app.mount();
+  
+      await app.mount({
+        dom: appRef.current
+      });
+
+      return () => {
+        app.unmount();
+      };
     })()
-  })
+  });
+
   return <div ref={appRef}/>
 }