+++ packages/core/kernal/src/index.ts
@@ -1,15 +1,26 @@
 import { Application } from './application/Application';
-import { createEventBus } from './application/createEventBus';
 
 // @ts-ignore
 window.__CONSOLE_OS_GLOBAL_VARS_ = {};
 
-export { mountApp, isAppRegistered, start, createMicroApp, unmount, mount, load, distroy, getExposedModule, loadExposedModule, update } from './mount';
+export { 
+  mountApp, 
+  isAppRegistered, 
+  start, 
+  createMicroApp,
+  unmount,
+  mount,
+  load,
+  destroy,
+  update,
+  getExposedModule,
+  loadExposedModule,
+} from './mount';
 
-export { createEventBus }
+export { createEventBus } from './application/createEventBus';
 
 export { prefetch } from './prefetch';
 
-export { SandBoxOption } from './type';
+export { SandBoxOption, AppInfo } from './type';
 
 export type OSApplication = Application;